# Serve endpoint with hypercorn and uvloop