--- a/setup.cfg
+++ b/setup.cfg
@@ -21,6 +21,7 @@ install_requires =
     aiortc>=1.3.2
     cloudpickle>=1.6.0
     globus-sdk>=3.3.0
+    hypercorn[uvloop]>=0.13.0
     lazy-object-proxy>=1.6.0
     parsl>=1.0.0
     quart>=0.18.0