# Description
<!--- Describe your changes in detail --->

Use Hypercorn with uvloop (rather than Python's native event loop) for serving the endpoint.

#### Before:

| Clients          | 1      | 4      | 8     |
|------------------|--------|--------|-------|
| Avg Latency (ms) | 1.238  | 3.042  | 5.780 |
| QPS              | 807    | 1315   | 1384  |

#### After:

| Clients          | 1     | 4     | 8     |
|------------------|-------|-------|-------|
| Avg Latency (ms) | 1.144 | 1.917 | 3.892 |
| QPS              | 873   | 2086  | 2056  |

Test was done using the QPS benchmark in [proxystore-benchmarks](https://github.com/proxystore/proxystore). The endpoint and clients were hosted on the same physical machine and the `/endpoint` route was queried.

#### Future changes:

This change sets us up for using multiple workers if needed for serving the endpoint. The only limiting factor is that the data storage in the endpoint also needs to support multiple independent processes (SQLite, Redis, etc.).

### Fixes
<!--- List any issue numbers above that this PR addresses --->

- Fixes #72 
- Fixes logging regression in #69 

### Type of Change
<!--- Check which off the following types describe this PR --->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation update (no changes to the code)
- [ ] CI change (changes to CI workflows, packages, templates, etc.)

## Testing
<!--- Please describe the test ran to verify changes --->

Updated unit tests where needed and ran QPS benchmark to verify performance improvements.

## Pull Request Checklist

Please confirm the PR meets the following requirements.
- [x] Code changes pass `pre-commit` (e.g., black, flake8, mypy, etc.).
- [x] Tests have been added to show the fix is effective or that the new feature works.
- [x] New and existing unit tests pass locally with the changes.
- [x] Docs have been updated and reviewed if relevant.